sample_project/src/config.py
from pathlib import Path
from dotenv import load_dotenv

# project configuration from .env (secret part)
env_path = Path(__file__).resolve().parent / '.env'
load_dotenv(dotenv_path=env_path)  # loads into os.environ

# project configuration
DATA_DIR = "../data"
RESULTS_DIR = "../results"

# data sources configuration
IRIS_URL = 'https://gist.githubusercontent.com/curran/a08a1080b88344b0c8a7/raw/0e7a9b0a5d22642a06d3d5b9bcbad9890c8ee534/iris.csv'
TITANIC_DATASET_SLUG = 'yasserh/titanic-dataset'
WIKI_LARGEST_COMPANIES = 'https://en.wikipedia.org/wiki/List_of_largest_companies_by_revenue'

