sample_project/requirements.txt
numpy
pandas
scikit-learn
kaggle
kagglehub
python-dotenv